+++ typography-frontend/package.json
@@ -8,6 +8,7 @@
     "@testing-library/user-event": "^12.6.2",
     "axios": "^0.21.1",
     "bootstrap": "^4.6.0",
+    "eslint": "^7.28.0",
     "formik": "^2.2.6",
     "json-server": "^0.16.3",
     "react": "^17.0.1",
@@ -42,5 +43,8 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "eslint-plugin-react": "^7.24.0"
   }
 }
