+++ typography-frontend/src/App.js
@@ -1,27 +1,26 @@
 import React from 'react'
 
-import {Redirect, Route, Switch} from 'react-router-dom'
-import BrowserRouter from 'react-router-dom/BrowserRouter'
+import { Route, Switch } from 'react-router-dom'
 
-import {Header, FeedbackSuccess, SignUp} from "./components";
-import {FeedbackPage, MainPage} from './pages/index'
-import NewsPage from "./pages/NewsPage";
-import NotFoundPage from "./components/NotFoundPage";
+import { Header, FeedbackSuccess, SignUp } from "./components";
+import { FeedbackPage, MainPage, CabinetPage, NewsPage } from './pages/index';
+import {NotFoundPage} from './common'
 
 function App() {
 
     return (
-        <BrowserRouter>
+        <>
             <Header/>
             <Switch>
-                <Route path='/' exact component={MainPage}/>
-                <Route path='/feedback' exact component={FeedbackPage}/>
-                <Route path='/feedbackSuccess' exact component={FeedbackSuccess}/>
-                <Route path='/news' exact component={NewsPage}/>
-                <Route path='/signUp' exact component={SignUp}/>
-                <Route component={NotFoundPage}/>
+                <Route path='/' exact component={MainPage} />
+                <Route path='/feedback' exact component={FeedbackPage} />
+                <Route path='/feedbackSuccess' exact component={FeedbackSuccess} />
+                <Route path='/news' exact component={NewsPage} />
+                <Route path='/signUp' exact component={SignUp} />
+                <Route path='/cabinet' exact component={CabinetPage} />
+                <Route component={NotFoundPage} />
             </Switch>
-        </BrowserRouter>
+        </>
     )
 }
