+++ typography-frontend/src/common/index.js
@@ -1,0 +1,7 @@
+import Header from "./Header/Header";
+import NotFoundPage from "./NotFoundPage/NotFoundPage";
+
+export {
+    Header,
+    NotFoundPage
+}
