+++ typography-frontend/src/components/Auth/SignUp/SignUp.jsx
