+++ typography-frontend/src/components/Auth/index.js
@@ -1,0 +1,5 @@
+import SignUp from "./SignUp/SignUp";
+
+export {
+    SignUp
+}
