+++ typography-frontend/src/components/Cabinet/Cabinet.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Cabinet = () => {
+    return (
+        <section>
+
+        </section>
+    );
+};
+
+export default Cabinet;
