+++ typography-frontend/src/components/Feedback/index.js
@@ -1,0 +1,7 @@
+import FeedbackContainer from "./FeedbackContent";
+import FeedbackSuccess from "./FeedbackSuccess";
+
+export  {
+    FeedbackContainer,
+    FeedbackSuccess
+}
