+++ typography-frontend/src/components/Main/main.css
