+++ typography-frontend/src/index.js
@@ -3,11 +3,6 @@ import ReactDOM from 'react-dom';
 
 import BrowserRouter from 'react-router-dom/BrowserRouter'
 
-import './css/index.css';
-import './css/style.css';
-import './css/feedback.css';
-import './css/sign.css';
-
 import App from './App';
 
 ReactDOM.render(
