+++ typography-frontend/src/pages/CabinetPage.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import {Cabinet} from "../components";
+
+const CabinetPage = () => {
+    return (
+        <Cabinet />
+    );
+};
+
+export default CabinetPage;
