+++ typography-frontend/src/pages/FeedbackPage.jsx
@@ -1,10 +1,10 @@
 import React from 'react';
 
-import {FeedbackContent} from '../components/index'
+import {FeedbackContainer} from "../components/Feedback";
 
 const FeedbackPage = () => {
     return (
-        <FeedbackContent/>
+        <FeedbackContainer/>
     );
 };
