+++ typography-frontend/src/pages/NewsPage.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import News from "../components/News";
+import News from "../components/News/News";
 
 const NewsPage = () => {
