+++ typography-frontend/src/pages/index.js
@@ -1,9 +1,11 @@
 import MainPage from "./MainPage";
 import FeedbackPage from "./FeedbackPage";
 import NewsPage from "./NewsPage";
+import CabinetPage from "./CabinetPage";
 
 export {
     MainPage,
     FeedbackPage,
-    NewsPage
+    NewsPage,
+    CabinetPage
 }
